Remove help text from errors; handle launching Proton Experimental
It is bad practice to spew the entire help if a flag is unrecognised. Further, because of the way proton-launcher.sh looks for $protonver, trying to run Proton Experimental will not work unless you invoke it as "- Experimental". Hacky change here checks if the user is asking for Proton Experimental and launches it explicitly.